toyota_diagnostic_pro/requirements.txt
streamlit
obd
python-can
cantools
rich
pandas
plotly
reportlab
openpyxl
xlsxwriter
